Upload images route fixed
