+++ src/app/api/uploadimg/route.ts
@@ -1,13 +1,6 @@
 import { mongooseConnect } from "@/lib/mongoose";
 import { NextResponse } from "next/server";
-import multiparty from "multiparty";
 import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
-import fs from "fs";
-import mime from "mime-types";
-import { IncomingMessage, request } from "http";
-import { NextApiRequest } from "next";
-var toArray = require('stream-to-array')
-
 
 export const config = {
   api: {
@@ -22,18 +15,17 @@ export async function POST(req: Request) {
   const bucketName = process.env.AWS_BUCKET_NAME;
   const fileName = process.env.AWS_FILE_NAME;
   const formData = await req.formData();
-  console.log(formData)
-  const file = formData.get('file');
+  const file = formData.get("file");
 
   if (!accessKeyId || !secretAccessKey) {
-    throw new Error('AWS credentials Error');
+    throw new Error("AWS credentials Error");
   }
 
   const client = new S3Client({
-    region: 'us-east-1',
+    region: "us-east-1",
     credentials: {
       accessKeyId,
-      secretAccessKey
+      secretAccessKey,
     },
   });
 
@@ -42,8 +34,6 @@ export async function POST(req: Request) {
     if (file instanceof Blob) {
       const ext = file.name.split(".").pop();
       const newFileName = Date.now() + "." + ext;
-
-      // Convert Blob to stream
       const fileStream = file.stream();
       const reader = fileStream.getReader();
       const chunks = [];
@@ -57,24 +47,22 @@ export async function POST(req: Request) {
       }
 
       const buffer = Buffer.concat(chunks);
-      console.log(buffer)
 
-      await client.send(new PutObjectCommand({
-        Bucket: bucketName,
-        Key: newFileName,
-        Body: buffer, // use the file stream instead of reading from file system
-        ACL: "public-read",
-        ContentType: file.type || undefined, // use the type property of the File object
-      }));
-
+      await client.send(
+        new PutObjectCommand({
+          Bucket: bucketName,
+          Key: newFileName,
+          Body: buffer, // use the file stream instead of reading from file system
+          ACL: "public-read",
+          ContentType: file.type || undefined, // use the type property of the File object
+        })
+      );
 
       const link = `https://${fileName}.s3.amazonaws.com/${newFileName}`;
       links.push(link);
     }
 
-
     return NextResponse.json({ links });
-
   } catch (error) {
     return NextResponse.json({ message: "Internal Server Error", error });
   }
